Classes can access `protected static` members of their Java superclasses.
This more-or-less reverts the fix added in 1ebbe029dd7ba. Instead
of wantonly treating `protected static` Java members like they're
public, it suffices to modify `isSubClassOrCompanion` to consider
a class a "subclass or companion" of the companions of its parent
classes, provided the parent is a Java-defined class.

It's a bit wacky, but comparing the Java and Scala definitions of
`protected` access:

SLS 5.2.8:
    Protected members of a class can be accessed from within

    - the template of the defining class,
    - all templates that have the defining class as a base class,
    - the companion module of any of those classes.

JLS 6.6.1:
    A member (class, interface, field, or method) of a reference type
    ... is accessible only if the type is accessible and the member
    or constructor is declared to permit access:

    ...
    - if the member or constructor is declared protected, then access
      is permitted only when one of the following is true:

      - Access to the member or constructor occurs from within the
        package containing the class in which the protected member
        or constructor is declared.
      - Access is correct as described in §6.6.2.

JLS 6.6.2.1:
    Let C be the class in which a protected member is declared.
    Access is permitted only within the body of a subclass S of C.

The important differences are that Java protected members are
indeed accessible from the package that they're declared in (this
behavior is implemented by `accessBoundary`, and that `static`
methods are treated the same as instance methods for purposes of
access control. Therefore, I feel comfortable adding the third
case to `isSubClassOrCompanion`, thus taking the Java access
control model when dealing with Java-defined types.

I also scouted a grammar tweak, which makes this have a few more
files in the diff than would be expected.

Fixes scala/bug#6934, review by adriaanm and (of course) whoever
feels like swinging by and saying hi.