Use AnyRefMap in picker, we don't actually need a linked map
AnyRefMap is quite a bit faster as it avoids cooperative equality and
doesn't maintain the linked list of entries in insertion order.

The linked-ness of this particular map dates back to everone's favourite commit,
"Massive check-in for IDE." a205b6b0. I don't see any motivation for it,
as the contents are never iterated.

Given that a number of maps were made linked in that commit, I conclude
that this one was done speculatively while developing that change,
but wasn't needed.