Avoid an AST traversal in the erasure phase
A separate traversal between preErase and re-typechecking was
responsible for filling in super-init calls in trait constructors
and adjusting the template parents in the AST to align with the
info-transformed class info.

This commit removes this traversal and handles those concerns in
the typechecking traversal

This was probably harder to do before the changes to the trait encoding
in Scala 2.12.