Simplify skipping of private methods in the classfile parser
In Scala 2.11 under -optimize, symbols were created for private methods
in the classpath to support the old inliner, which used to parse bytecode
of methods (which could refer to private methods) into the ICode IR.

The new backend deals exclusively in bytecode via ASM, and doesn't need any
help from the classfile parser or symbol table.

`settings.optimise` is already false when using the non-deprecated way of
enabling the new optimizer, `-opt`.