@implicitNotFound messages can access type args from prefixes
Motivation:

```scala
class HMap[Rel[_, _]] {
  def get[K, V](k: K)(implicit rel: Rel[K, V]): V
  def add[K, V](k: K, v: V)(implicit rel: Rel[K, V]): HMap[Rel]
}

trait ListOf[Upper] {
  @implicitNotFound("Cannot add (${K}, ${V}) to a HMap (is ${V} =:= List[${K}] and <: ${Upper}?)")
  trait Lambda[K, V]
}
object ListOf {
  implicit def rel[Upper, T <: Upper]: ListOf[Upper]#Lambda[T,List[T]] = ...
}
```

and various other type-level weirdnesses.

The implementation is pretty simple: traverse the type, grabbing
`(sym.tparams zip args)` for each `TypeRef` we can get our hands
on. For `validate`, of course, we have the symbol of the type
definition to go off of, and just grab the `tparams` from all of
the types in the `ownerChain`.