Replace dummy `unapply-selector` by ref to real val
Once a pattern match is expanded, we have a `val` in hand
that we can use to represent the input to the unapply method.
We already spliced that `val` into the call to the extractor,
but neglected to substitute in the types. For `unapply`s
with dependent result types, that left a bunch of types
containing `<unapply-selector>.type` in the trees. No more!

The substitution is performed in three ways:
  - when splicing the argument (as mentioned above);
  - when determining the types of binders for subpatterns
    (this area was cleaned up in parent commit)
  - an additional SubstTreeMaker chained before the
    extractor treemaker (this one is for good measure;
    we could actually see if we truly need it)