Drop useless failing assert in `mkCast`
I added it in e1c732d to encode an assumption, but it's
not an invariant that must be enforced (erasure will erase anyway).

More generally / longer term, we should reevaluate the use of `dealias`
for types that are meant to approximate their equivalent in the erased
type system.

Specifically, for `mkCastPreservingAnnotations` I think we should
either `normalize` or do nothing, but the half measure of `dealias`
does not make sense, as the logic behind a cast operates on the fully
normalized type, not just on a dealiased type (think refinements with
type aliases). It would be ok to do nothing here, because erasure will
convert the type to something that can be casted anyway.