Clean the original tree attachment in erasure
The original tree attachment is useful for compiler plugins and macros.
However, this commit constraints compiler plugins to run before erasure
if they want to inspect the original trees so that we free up memory as
soon as possible.

This commit removes the attachment in pre erasure, taking advantage that
we modified literals before.