Massage performance with statistics optimizations
The following commit applies some minor optimizations to the statistics
code and adds a heuristic to make sure that impact of runtime is as
minimal as possible.

These changes are motivated by a recorded performance degradation of
around 2-3% when statistics are not stored in an object with static
values.

The idea of the heuristic is to have a filter that is true almost all
the times, with a static `areSomeColdStatsEnabled`. This static will be
true whenever a `Global` instance has enabled statistics. If it is true,
then the statistics infrastructure will check if the actual global
should record statistics, or it's another global instance the one that
enabled them.

Therefore, when enabling statistics in one global, we'll pay an overall
performance degradation of 2/3% for all globals of a given classloaded
scalac + the statistics overhead in the global that wants to record
statistics.