Return in PartialFun relative to right enclosing method
Where the right enclosing method is the lexically enclosing one,
in the code as written by the user, not the one the body actually
ends up being type checked in (`applyOrElse`).

Since the current owner impacts more than just type checking
`Return` trees, we adjust the currently logically enclosing method
as we do for lazy getters (that adjustment happens in Context,
but retronym pointed out we can do it closer to where it's needed
in this case -- TODO: can we do the same for lazy vals).