Hot fix registration of timers and counters
Because of initialization order of the cake, `statistics` is initialized
after the timers/counters in the cake are, so when it hits the
constructor of `Quantity` those are not registered in `qs`.

This meant that even though those objects were initialized, statistics
were not reported. This change hot fixes it so that they are indeed
reported. It does so by removing the guard that checked whether
statistics were enabled. From now on, for simplicity and correctness, we
will always register timers and counters that are initialized. This
should have no impact in performance, since it's done only once when
everything is initialized, and it's just an addition to a map.