Backport suport for Java 9 platform classpath
In Scala 2.13.x, we've supported this by using the NIO
jrt:// virtual filesystem. We can't (easily) ship code
in older Scala distributions that require Java 7+, but
we can fall back to use of `javax.tools`, which offers
programatic access to the Java compiler's classpath
implementation.

This is incompatible with the performance trick used in
the runner scripts that places the Scala compiler on the JVM
boot classpath to avoid the cost of classfile verification.
The boot classpath can only see the base platform on Java 9,
which puts `javax.tools` out of access.

But, we can use `scalac -nobootcp` to make things work!
This obviates the workaround used in SBT which exports
rt.jar from JDK9 and passes to scalac as a custom boot
classpath.