Don't assume that class literals of value classes wrap TypeRefs.
When erasure comes around to munge on a TypeRef to a derived
value class, it swaps it out for an ErasedValueType that will
get posterased to the underlying type. This behavior is wholly
inappropriate for class literals, which one would prefer remain
referencing the derived value class itself, and so such TypeRefs
are rightfully exempted from the specialScalaErasure.

However, TypeRefs aren't the only way to reference derived value
classes: the type may be an ExistentialType or some other wrapper
of a TypeRef to a derived value class. The easy way to check is to
call typeSymbol.isDerivedValueClass, and use the javaErasure (which
erases derived value class TypeRefs the same as normal TypeRefs).

There are three cases where this matters:
  - erasing classOf during erasure (scala/bug#10551)
  - erasing the element type of ArrayValues during erasure (scala/bug#10646)
  - erasing classOf inside Java annotations during jvm

fixes scala/bug#10551; fixes scala/bug#10646 (not like github honors this these days)