Allow singleton group annotations in Scaladoc, fixes scala/bug#10078
In the special case that every member of a package used the same group
annotation, Scaladoc would ignore the group. This was due to filtering
that checked the number of extant groups (including the default
"Ungrouped" group), skipping groups if only one group was found.
Unfortunately, the "Ungrouped" group could be missing if all templates
used the same group!

Fix this by filtering out the "Ungrouped" group entirely, and checking
for the existence of any other groups. Make the solution a bit less
fragile by referencing the default group constant defined in
ModelFactory, rather than the default group's string directly.