Remove compiler dependency on `@remote`
The `@remote` annotation will be removed in 2.13.
It was deprecated in 2.12, and will be gone entirely in 2.13.0-M4,
but first we must release a milestone that no longer refers to it
(so that the compiler can bootstrap without a definition for `@remote`).

This cuts tie 1 of 3 to the Java profiles beyond compact1.