Don't emit the serialVersionUID field on traits.
These days, it's `private`, and the JVM doesn't like private
fields on interfaces. Since the field doesn't have an effect
when put on an interface, we might as well just not.

Also fix the capitalization of the annotation in the warning,
and also warn when it's used on abstract classes (where it's
also useless).

Fixes scala/bug#10610; references scala/collection-strawman#296.