Remove @SerialVersionUID from traits.
...and put them on the subclasses, out of an abundance of caution.

Starting in 2.13.0-M3, the `serialVersionUID` field is emitted
as `private`, and so won't be applied to `trait`s, which can't
have `private` members. The value didn't have an effect on the
Java serialization algorithm, so this changes nothing.

Also, compute the serialVersionUID value for the subtypes of
those traits, since evidently it was intended that they have
a stable value.

Fixes #296.