Import selector respects backquotes
To disambiguate backquoted underscore, the convention
for ImportSelector is tweaked by using EMPTY instead of
WILDCARD as the special value. So EMPTY => EMPTY is
wildcard import, nonempty => EMPTY is hidden.

This encoding also avoids a null value for the rename field.

A rename is detected more simply by name != rename.

Underscore is not an identifier

Don't take `val _: T` as defining a member or variable
with that name.

Tweak encoding of import selector

Use (WILDCARD, null) for wildcard as now,
and (WILDCARD, other) for backticked underscore,
which may be a rename.

Introduce some helpers on ImportSelector
to keep it straight.

Fix paradise