Enforce maxerrs more robustly
`-Xmaxerrs` is currently advisory and depends on the good will
of the reporter. The standard console reporter is well-advised,
but others are not, such as sbt, which uses an sbt setting to
control logging.

This commit adds a simple filter in front of any reporter
deemed untrustworthy.

Comments added as reminder that `reporters.Reporter` is
obsolete and implementations will use `internal.Reporter`.