Updated isInstanceOf[Singleton] to match SIP decision.
isInstanceOf is supposed to be a dynamic test and at run time all
non-null values are dynamically singletons, so warn of a fruitless type
test and compile as true for subtypes of AnyVal and as equivalent to ne
null for subtypes of AnyRef. See discussion on the related Dotty
PR: https://github.com/lampepfl/dotty/pull/3681.