Make Typer#dyna.mkInvoke more accurate
When searching the context for the tree in question, make sure to
actually check that the tree under scrutiny matches the one we're
looking for. This means that the check for varargs won't give a false
positive if the context is too large. Fixes scala/bug#7420.

Adjust the "vararg not supported" error so it doesn't hide other errors.