Add the sm"" interpolator to the standard library.
This has been a feature available to the compiler since
a0001fcfd (back in 2012!), but somehow never made it to
the masses. It's useful, so let's put it in `StringContext`
so all can enjoy its margin-stripping goodness.

Rather than just copy the implementation over, I've taken
the liberty of turning it into a macro. Since there are
plans to intrinsify the `s""` and `raw""` interpolators
in the near future, the macro emits a call to `raw""` in
the form that will be expected by that transformation.

Since the new interpolator can't yet be used in reflect/compiler,
there's a `private[scala]` shim `sm0` for their benefit. After
reSTARR, I'll change them back to the stdlib one.