Fix overzealous protected access check for Java static members.
Both Java and Scala require that, to access a protected member
of a class `C`, the access must occur inside a class `S` which
extends `C`. Moreover, the type of the qualifier needs to be a
subclass of `S`. However, when a Java `static` member is being
selected, Java doesn't care about the prefix (there is no such
concept in Java-land). In Scala, however, the selection occurs
from a fictional companion module made to house all the static
members, and it's not likely to have any subclass relationship
with the classes that we care about.

Therefore, when selecting from Java-defined modules,
ignore the prefix check.

This worked before 01c3bbb9c, which tightened up the restriction
a little too far.

Fixes scala/bug#10568,
and fixes scala/bug#10597,
but does not fix my dignity.