Speed up creation of [Abs]TypeErrors.
`Throwable#fillInStackTrace` can be expensive, so minimize
the number of calls we make to it.

- `AbsTypeError` and its subclasses are never thrown, so
  they don't need to extend `Throwable` at all.
- `TypeError`s are thrown, but if the user sees them it's
  a compiler bug anyhow, so only populate the stack trace
  if we're in `-Ydebug` mode.

This also adds a minute bit of clarity to the distinction
between `TypeError` and `AbsTypeError`: you can't throw
the latter.

Contribution note: retronym did this independently last
October, because all good ideas are already had. Found
this out when asking for permission on gitter rather than
forgiveness on github. Still committing, though, because
I've noticed that smaller changes tend to get merged sooner.