Avoid accumulating ill-kinded bounds during unification
Turning partial unification on unconditionally revealed that in some
error cases it was possible for the partial unification branch of
unifySpecific to add inappropriately kinded bounds to the type variable
being solved. These later blow up in matchingBounds in GlbLubs where the
mismatched kinds result in a NoCommonType exception being thrown which
crashes the compiler reporting "lub/glb of incompatible types".

This commit prevents that from happening by checking that the added
bounds will match the kind of the TypeVar they're being added to. If
they don't then the bound is not added and unification fails.