Cache classloaders for compiler plugins
When users compile their sources with external compiler plugins whose
classes are not found in the compiler's classpath, the compiler needs to
classload them every time they want to be used.

This process can slow down compilation severely, as shown in
https://github.com/scala/scala-dev/issues/458. This cost is due to the
JVM JIT-compiling the recently loaded classes, and the JIT deoptimizing
because new classes violate the optimization assumptions taken via Class
Hierarchy Analysis (CHA).

The cost of dynamic classloading can be mitigated by caching the
classloaders for compiler plugins based on their file stamps, in a
similar way to how the compiler currently caches classpaths. If the
file stamps change, the compiler plugin will be loaded again. If they
don't (which is by far the most common scenario), the classloaders will
be reused.

Fixes scala/scala-dev#458. Thanks to Jason for suggesting the fix.