Improve error messages for method references
A method is not a value, and thus must somehow be converted
to an expression -- unless we're specifically looking for a method.

We can either insert an application (to implicit args or empty arg list),
or lift to a function by eta-expansion. If those are not possible,
we report an error (unstable tree -- could be refined further).

Tests for scala/bug#10474, scala/bug#10731