Don't copy the OVERRIDE flag in refinement decls
typedRefinement defers the setting of this flag until the end
of the compilation unit, which means that inferred types that are
derived from the written refinement type can be unstable depending
on whether they were computed before or after the flag was set.

An alternative fix might be to just remove the setting of OVERRIDE
in typedRefinement.unitToCheck.

(cherry picked from commit f6ca3dd53e7937a6750266625f3f47e28a0b38a2)