Eta-expand methods passed as functions to overloaded methods
The stricter automatic eta-expansion of method types to function
types in 2.13 interacts badly with the improved overload resolution
for higher-order methods. When a method reference is passed to such a
method without explicit eta-expansion, the shape-based lubbing
is not applied. Therefore the method is checked without an expected
type and does not eta-expand as it would when passed to a
non-overloaded method.

This is an attempt to fix the problem by using lubs more aggressively.
Unlike function and pattern-matching function literals method references
cannot be detected syntactically as early as we would like (as they are
only Idents without a symbol or type at that point) so we try to lub
all argument types and use that for typing the alternatives unless a
typed function literal is passed (because that might not typecheck
against the lub).