Fix lookup of default getter in scope
By slightly modifying an existing test to force creation of
default getters for both `bar` methods _before_ typechecking the
application, I was able to show a latent bug in the way the default
getter is looked up in scope.

The bespoke `Context.lookup` method did not respect shadowing,
but rather considered the two, same-named default getters as
overloaded. Because the overloaded symbol had NoSymbol as its
owner, which didn't match the expected owner, neither default
was eligible.

This commit brings the code more into line with `Context.lookupSymbol`
and respects shadowing.