drop -Yuse-stupid-types
this was added by Gilles in 2010 "for internal use" and with the
description "Print the types of inherited members as seen from their
original definition context. Hint: you don't want to do that!"

so, it isn't clear what this was even for originally. but, googling and
searching GitHub turns up no uses.