Refactor typechecking of array instantiation for performance
Defining the extractor object `ArrayInstantation` in `typed1` meant
that we needed to thread a `LazyRef` through to `typedApply`
(even if we were dealing with a different AST node!).

This commit moves the extractor to `TreeInfo` after extracting the
core part that relies on `Typer.this`.

(cherry picked from commit d1c90ec3dec2bf72546edcda9e4f696882a7ec61)