Preserve order of decls through pickle/unpickle
The pickle format does not explicitly enccode the order of
decls. Instead, symbols are entered into an index in the
order that they are found by the pickler, either as a definition
or as a reference. During unpickling, symbols are read and entered
into the owner's decls in that order.

This is a cause of unstable compiler output: a class that mixes
in members from some trait will have a different order if it is
compiled jointly with / separately from that trait.

This commit modifies the pickler with an initial pass that reserves
index entries for all declarations in the declaration order.

The pickle format and the unpickler are unchanged.