Warn also on mirror class clobbering.
... and mention the source file if it's different.

This confused someone on gitter today, and I noticed that the clobbering
check misses out on `class Foo; object foo` (whereby `foo`'s mirror will
overwrite `Foo`, on a Mac). Slap a call to the check in `genMirrorClass`,
and folks are duly warned.

Also drop `neg/case-collision2`, which existed just to test the same
test on both JVM backends.