First-class treatment of ClassTag-based factories
This reinstates the factory part of the abstraction for arbitrary
evidence type constructors which we previously abandoned in favor of
SortedIterableFactory. Now we have a good reason for an
EvidenceIterableFactory with several Ordering-based and several
ClassTag-based factory instances.

ImmutableArray and WrappedArray now have a ClassTag-based factory in
their companion object so that building unboxed arrays with a proper
element type is always the default (and used automatically in calls like
`xs.to(ImmutableArray)` which would previously go through the regular
IterableFactory and build a boxed representation). IterableFactories are
still available as `{ImmutableArray/WrappedArray}.untagged`.

There are no new abstractions at the collection level. We could
consider reintroducing a ConstrainedIterable but at lower levels of
the hierarchy it doesn’t make sense because the Ordering- and
ClassTag-based collections are of very different types.

There are also some fixes for bad arraycopy calls (between boxed and
unboxed representations) and unnecessary boxing in combinators in
ImmutableArray.