Avoid IIOBE in phase stack
Flatten's info transform ends calls `enteringMixin(enteringFlaten(...))`
all the way up the prefix, which violate my assumption that 128 recursive
atPhase calls should enough for anyone when we encounter heavily nested
code (such as that regrettably still emitted by the REPL when many imports
precede the current line).

This commit switches to using an ArrayBuffer that will grow as needed,
and also disables the entire maintenance of the phase stack when logging
is disabled.