Stabilizing definitions fixes
https://github.com/scala/scala/pull/5999 introduces a synthetic val for
the unstable LHS of an implicit method application where the method type
depends on the LHS. Unfortunately this only handles the case where an
expression requires a _single_ stabilizing val: if the expression
involves chained method applications, each requiring a stabilizing val,
then only the last of these is emitted resulting in a failure during
bytecode generation,

object Test {
  class Bar {
    class Foo {
      type Out
    }
    object Foo {
      implicit def foo: Foo { type Out = Bar } = ???
    }

    def foo(implicit foo: Foo): foo.Out = ???
  }

  (new Bar).foo.foo
}

results in a NoSuchElementException,

java.util.NoSuchElementException: value stabilizer$1
        at scala.collection.mutable.AnyRefMap$ExceptionDefault.apply(AnyRefMap.scala:425)
        at scala.collection.mutable.AnyRefMap$ExceptionDefault.apply(AnyRefMap.scala:424)
        at scala.collection.mutable.AnyRefMap.apply(AnyRefMap.scala:180)
        at scala.tools.nsc.backend.jvm.BCodeSkelBuilder$PlainSkelBuilder$locals$.load(BCodeSkelBuilder.scala:392)
...

Prior to this commit only a single stabilizer is recorded on the
expression tree which is overwritten by later ones. This is fixed here
by allowing multiple stabilizers to be recorded and emitted.

In addition to that, previously spurious dependencies were picked up due
to insufficient dealiasing. The following,

object Test {
  class Bar {
    class Foo {
      type Out
    }
    object Foo {
      implicit def foo: Foo { type Out = Bar } = ???
    }

    def foo(implicit foo: Foo): foo.Out = ???
  }

  (new Bar).foo.foo
}

triggered the preceding error despite not having a true dependency which
would require a stabilizer in the first place. The test for a dependent
argument now dealiases to eliminate these spurious dependencies.

Finally we avoid retyping stabilizing definitions when they are emitted,
fixing https://github.com/scala/bug/issues/10736.