Issue macro expansion errors during "late expansion"
The typechecker defers macro macro expansion while the macro
application has type arguments that are yet to be determined.
Instead, it proceed with typechecking the surrounding
expression, and then performs another pass on the resulting tree
when the type inference has fixed the type parameters.

This "late expansion" typechecks the macro application with a
typer focussed on the original Context of the application,
and any errors issued by the macro (either explicit c.error
or c.abort, or an exception) are issued to that contexts's
reporter.

However, if that reporter was setup to buffer errors, rather
than immediately report them, these macro expansion errors
would sit in the never be issued, and the unexpanded macro
application would remain in the tree, doomed to be flagged with
a "macro not expanded" error in refchecks.

This commit copies any buffered errors to the currently active
typer context reporter after late expansion.

I refactored the existing code that did this to make it more
easily reusable in this context.